Found that the logic for the navigation screen boxes was not taking in to account that the color set at the beginning was not stored for later.
* Added property to store the color for each NavigationBox.
* Moved the Models into their respective NavigationBoxControls.
* Fixed Responses to only appear when items have been added.
