+++ Obiddable.Win/Navigation/ItemNavigationBoxControl.cs
@@ -1,5 +1,6 @@
-﻿using Obiddable.Win.Library;
-using Obiddable.Win.Navigation;
+﻿using Obiddable.Library.Bidding;
+using Obiddable.Library.Cataloging;
+using Obiddable.Win.Library;
 
 namespace Obiddable.Win.UI.Bidding.Navigation;
 
@@ -10,7 +11,7 @@ public partial class ItemNavigationBoxControl : BidNavigationBoxControl
         InitializeComponent();
         SetClickEventOnControls(this);
         SetTitle("Items");
-        SetButtonColor(ApplicationColors.Items);
+        EditButtonColor = ApplicationColors.Items;
     }
 
     protected override void InitLabels()
@@ -19,5 +20,19 @@ public partial class ItemNavigationBoxControl : BidNavigationBoxControl
         itemsCount.Text = boxModel.ItemsCount.ToString();
         categoriesCount.Text = boxModel.CategoriesCount.ToString();
         EditEnabled = boxModel.CanEditItems;
+    }
+
+    public class ItemBoxModel
+    {
+        public ItemBoxModel(Bid _bid)
+        {
+            ItemsCount = _bid.GetItemsCount();
+            CategoriesCount = _bid.GetItemCategoriesCount();
+            CanEditItems = _bid.CanEditItems();
+        }
+
+        public int ItemsCount { get; private set; }
+        public int CategoriesCount { get; private set; }
+        public bool CanEditItems { get; private set; }
     }
 }
