--- Obiddable.Win/Navigation/PurchaseOrderBoxModel.cs
@@ -1,19 +1,0 @@
-﻿using Obiddable.Library.Bidding;
-using Obiddable.Library.Purchasing;
-
-namespace Obiddable.Win.Navigation;
-
-public class PurchaseOrderBoxModel
-{
-    public PurchaseOrderBoxModel(Bid bid)
-    {
-        PurchaseOrders = bid.GetPurchaseOrdersCount();
-        PurchasedItems = bid.GetPurchaseOrdersItemsCount();
-        TotalPrice = bid.GetPurchaseOrdersTotalPriceSum();
-        CanEditPurchaseOrders = bid.CanEditPurchaseOrders();
-    }
-    public int PurchaseOrders { get; private set; }
-    public int PurchasedItems { get; private set; }
-    public decimal TotalPrice { get; private set; }
-    public bool CanEditPurchaseOrders { get; private set; }
-}
