+++ Obiddable.Win/Navigation/PurchaseOrderNavigationBoxControl.cs
@@ -1,5 +1,6 @@
-﻿using Obiddable.Win.Library;
-using Obiddable.Win.Navigation;
+﻿using Obiddable.Library.Bidding;
+using Obiddable.Library.Purchasing;
+using Obiddable.Win.Library;
 
 namespace Obiddable.Win.UI.Bidding.Navigation;
 
@@ -10,7 +11,7 @@ public partial class PurchaseOrderNavigationBoxControl : BidNavigationBoxControl
         InitializeComponent();
         SetClickEventOnControls(this);
         SetTitle("Purchasing");
-        SetButtonColor(ApplicationColors.Purchasing);
+        EditButtonColor = ApplicationColors.Purchasing;
     }
 
     protected override void InitLabels()
@@ -20,5 +21,21 @@ public partial class PurchaseOrderNavigationBoxControl : BidNavigationBoxControl
         purchasedItemsValue.Text = boxModel.PurchasedItems.ToString();
         totalPriceValue.Text = boxModel.TotalPrice.ToString("C");
         EditEnabled = boxModel.CanEditPurchaseOrders;
+    }
+
+    public class PurchaseOrderBoxModel
+    {
+        public PurchaseOrderBoxModel(Bid bid)
+        {
+            PurchaseOrders = bid.GetPurchaseOrdersCount();
+            PurchasedItems = bid.GetPurchaseOrdersItemsCount();
+            TotalPrice = bid.GetPurchaseOrdersTotalPriceSum();
+            CanEditPurchaseOrders = bid.CanEditPurchaseOrders();
+        }
+
+        public int PurchaseOrders { get; private set; }
+        public int PurchasedItems { get; private set; }
+        public decimal TotalPrice { get; private set; }
+        public bool CanEditPurchaseOrders { get; private set; }
     }
 }
