+++ Obiddable.Win/Navigation/RequestorNavigationControl.cs
@@ -1,5 +1,5 @@
-﻿using Obiddable.Win.Library;
-using Obiddable.Win.Navigation;
+﻿using Obiddable.Library.Bidding;
+using Obiddable.Win.Library;
 
 namespace Obiddable.Win.UI.Bidding.Navigation;
 
@@ -11,7 +11,7 @@ public partial class RequestorNavigationControl : BidNavigationBoxControl
         InitializeComponent();
         SetClickEventOnControls(this);
         SetTitle("Requesting");
-        SetButtonColor(ApplicationColors.Requesting);
+        EditButtonColor = ApplicationColors.Requesting;
     }
 
     protected override void InitLabels()
@@ -22,5 +22,23 @@ public partial class RequestorNavigationControl : BidNavigationBoxControl
         estimatedPriceValue.Text = boxModel.EstimatedPrice.ToString("C");
         estimatedPriceWithOverridesValue.Text = boxModel.EstimatedPriceWithOverrides.ToString("C");
         EditEnabled = boxModel.CanEditRequestors;
+    }
+
+    public class RequestorBoxModel
+    {
+        public RequestorBoxModel(Bid bid)
+        {
+            Requests = bid.GetRequestsCount();
+            Requestors = bid.GetRequestorsCount();
+            EstimatedPrice = bid.GetRequestorsEstimatedPrice();
+            EstimatedPriceWithOverrides = bid.GetRequestorsEstimatedPriceWithOverrides();
+            CanEditRequestors = bid.CanEditRequestors();
+        }
+
+        public int Requests { get; private set; }
+        public int Requestors { get; private set; }
+        public decimal EstimatedPrice { get; private set; }
+        public decimal EstimatedPriceWithOverrides { get; private set; }
+        public bool CanEditRequestors { get; private set; }
     }
 }
