--- Obiddable.Win/Navigation/VendorResponseBoxModel.cs
@@ -1,21 +1,0 @@
-﻿using Obiddable.Library.Bidding;
-using Obiddable.Library.Requesting;
-using Obiddable.Library.Responding;
-
-namespace Obiddable.Win.Navigation;
-
-class VendorResponseBoxModel
-{
-    public VendorResponseBoxModel(Bid bid, IRequestingRepo repo)
-    {
-        VendorResponses = bid.GetVendorResponsesCount();
-        ItemResponses = bid.GetVendorResponsesItemResponsesCount();
-        NoResponseItems = bid.GetNoResponseItemsCount(repo);
-        CanEditVendorResponses = bid.CanEditVendorResponses();
-    }
-
-    public int VendorResponses { get; private set; }
-    public int ItemResponses { get; private set; }
-    public int NoResponseItems { get; private set; }
-    public bool CanEditVendorResponses { get; private set; }
-}
