+++ Obiddable.Win/Navigation/VendorResponseNavigationBoxControl.cs
@@ -1,19 +1,21 @@
-﻿using Obiddable.Library.EF.Requesting;
+﻿using Obiddable.Library.Bidding;
+using Obiddable.Library.EF.Requesting;
 using Obiddable.Library.Requesting;
+using Obiddable.Library.Responding;
 using Obiddable.Win.Library;
-using Obiddable.Win.Navigation;
 
 namespace Obiddable.Win.UI.Bidding.Navigation;
 
 public partial class VendorResponseNavigationBoxControl : BidNavigationBoxControl
 {
     private readonly IRequestingRepo _requestItemsRepo = new EFRequestingRepo();
+
     public VendorResponseNavigationBoxControl()
     {
         InitializeComponent();
         SetClickEventOnControls(this);
         SetTitle("Vendor Responses");
-        SetButtonColor(ApplicationColors.VendorResponses);
+        EditButtonColor = ApplicationColors.VendorResponses;
     }
 
     protected override void InitLabels()
@@ -23,5 +25,21 @@ public partial class VendorResponseNavigationBoxControl : BidNavigationBoxContro
         itemResponsesValue.Text = boxModel.ItemResponses.ToString();
         noResponseItemsValue.Text = boxModel.NoResponseItems.ToString();
         EditEnabled = boxModel.CanEditVendorResponses;
+    }
+
+    class VendorResponseBoxModel
+    {
+        public VendorResponseBoxModel(Bid bid, IRequestingRepo repo)
+        {
+            VendorResponses = bid.GetVendorResponsesCount();
+            ItemResponses = bid.GetVendorResponsesItemResponsesCount();
+            NoResponseItems = bid.GetNoResponseItemsCount(repo);
+            CanEditVendorResponses = bid.CanEditVendorResponses();
+        }
+
+        public int VendorResponses { get; private set; }
+        public int ItemResponses { get; private set; }
+        public int NoResponseItems { get; private set; }
+        public bool CanEditVendorResponses { get; private set; }
     }
 }
